Ammo crates now open the kit UI when pressing F, and can be punched to simply refill ammo. Also kit armory object support.

+++ UncreatedWarfare/Commands/Clear/ClearVehiclesCommand.cs
@@ -1,5 +1,4 @@
 using Uncreated.Warfare.Interaction.Commands;
-using Uncreated.Warfare.Logging;
 using Uncreated.Warfare.Translations;
 using Uncreated.Warfare.Vehicles;
 using Uncreated.Warfare.Vehicles.Spawners;
@@ -24,6 +23,11 @@ internal sealed class ClearVehiclesCommand : IExecutableCommand
 
     public async UniTask ExecuteAsync(CancellationToken token)
     {
+        if (Context.HasArgs(1))
+        {
+            throw Context.SendHelp();
+        }
+
         await _vehicleService.DeleteAllVehiclesAsync(token);
 
         foreach (VehicleSpawner spawner in _spawners.Spawners)
